fixed createcomment

+++ public/js/deletePost.js
@@ -1,29 +1,26 @@
 const deletePostHandler = async (e) => {
-  e.preventDefault();
+
+  const element = e.target;
   
-  const id = document.querySelector('.delete-btn').getAttribute('data-id');
+  if (element.matches(".delete-btn")) {
+    const id = element.dataset.id;
  
-  const res = await fetch(`/api/posts/${id}`, {
-    method: 'DELETE',
-    body: JSON.stringify({
-      id: id
-    }),
-    headers: {
-      'Content-Type': 'application/json'
+    const res = await fetch(`/api/posts/${id}`, {
+      method: 'DELETE',
+      body: JSON.stringify({
+        id: id
+      }),
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+
+    if (res.ok) {
+      document.location.replace('/dashboard/');
+    } else {
+      alert(res.statusText);
     }
-  });
-  
-  if (res.ok) {
-    document.location.replace('/dashboard/');
-  } else {
-    alert(res.statusText);
   }
 }
 
-const allDelBtns = document.querySelectorAll('.delete-btn');
-
-for (const btn of allDelBtns) {
-  btn.addEventListener('click', deletePostHandler);
-}
-
-// document.querySelector('.delete-btn').addEventListener('click', deletePostHandler);
+document.querySelector('.posts').addEventListener('click', deletePostHandler);
